# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Stealth Console API
  version: 0.1.0
  description: Control-plane API for Stealth Console. Browser sessions are opaque HttpOnly cookies and are never returned in JSON.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/organizations:
    get:
      operationId: listOrganizations
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Cursor' }]
      responses: { '200': { description: Organizations visible to the current account, content: { application/json: { schema: { $ref: '#/components/schemas/OrganizationsPage' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' } }
    post:
      operationId: createOrganization
      security: [{ sessionCookie: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateOrganizationRequest' } } } }
      responses: { '201': { description: Organization created; caller is owner, content: { application/json: { schema: { $ref: '#/components/schemas/OrganizationResponse' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '409': { $ref: '#/components/responses/Conflict' }, '413': { $ref: '#/components/responses/PayloadTooLarge' }, '415': { $ref: '#/components/responses/UnsupportedMediaType' }, '422': { $ref: '#/components/responses/ValidationError' } }
  /v1/organizations/{organizationID}:
    patch:
      operationId: updateOrganization
      description: Update an organization's display name and unique slug. Owners and admins may edit identity metadata; the organization ID is immutable.
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/OrganizationID' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateOrganizationRequest' } } } }
      responses: { '200': { description: Updated organization, content: { application/json: { schema: { $ref: '#/components/schemas/OrganizationResponse' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' }, '413': { $ref: '#/components/responses/PayloadTooLarge' }, '415': { $ref: '#/components/responses/UnsupportedMediaType' }, '422': { $ref: '#/components/responses/ValidationError' }, '500': { $ref: '#/components/responses/InternalError' } }
  /v1/organizations/{organizationID}/plan:
    get:
      operationId: getOrganizationPlan
      description: Return the effective organization plan, current period, server-owned resource limits, and bounded resource counts. Payment-provider identifiers are never returned.
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/OrganizationID' }]
      responses: { '200': { description: Organization plan and usage, content: { application/json: { schema: { $ref: '#/components/schemas/OrganizationPlanResponse' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '500': { $ref: '#/components/responses/InternalError' } }
...
  /v1/projects/{projectID}/users:
    get:
      operationId: listProjectUsers
      description: Members may read project application identities with a Console session. Server-side callers may use an X-Stealth-Key with users.read.
      security: [{ sessionCookie: [] }, { projectApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }, { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Cursor' }]
      responses:
        '200': { description: Application users visible to a project member, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectUsersPage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createProjectUser
      description: Owners/admins may create identities with a Console session. Server-side callers require an X-Stealth-Key with users.write.
      security: [{ sessionCookie: [] }, { projectApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateProjectUserRequest' } } } }
      responses:
        '201': { description: Application user created, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectUserResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/projects/{projectID}/users/{userID}:
    get:
      operationId: getProjectUser
      description: Members may read a project application identity with a Console session. Server-side callers may use an X-Stealth-Key with users.read.
      security: [{ sessionCookie: [] }, { projectApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }, { $ref: '#/components/parameters/UserID' }]
      responses:
        '200': { description: Application user, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectUserResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      operationId: deleteProjectUser
      description: Permanently delete a project application identity. Its sessions and recovery tokens are revoked atomically; server-side callers require an X-Stealth-Key with users.write.
      security: [{ sessionCookie: [] }, { projectApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }, { $ref: '#/components/parameters/UserID' }]
      responses:
        '204': { description: Application user deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/projects/{projectID}/users/{userID}/status:
    patch:
      operationId: updateProjectUserStatus
      description: Owners/admins may change status with a Console session. Server-side callers require an X-Stealth-Key with users.write.
      security: [{ sessionCookie: [] }, { projectApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }, { $ref: '#/components/parameters/UserID' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateProjectUserStatusRequest' } } } }
      responses:
        '200': { description: Application user with the requested status, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectUserResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: stealth_session }
    projectAppCookie: { type: apiKey, in: cookie, name: stealth_app_<project-id-without-hyphens>, description: Project-specific opaque HttpOnly cookie. The actual name is derived from the path projectID; it is never the Console stealth_session cookie. }
    projectApiKey: { type: apiKey, in: header, name: X-Stealth-Key, description: Project-bound server-to-server secret. Supported scopes include users, databases, storage, functions, sites, webhooks, realtime.read, and messaging read/write pairs; never send this header through the Console browser bridge. }
  parameters:
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Cursor: { name: cursor, in: query, schema: { type: string, format: uuid } }
    OrganizationID: { name: organizationID, in: path, required: true, schema: { type: string, format: uuid } }
...
    ProjectID: { name: projectID, in: path, required: true, schema: { type: string, format: uuid } }
...
    UserID: { name: userID, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    BadRequest: { description: Invalid JSON or parameter, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    Forbidden: { description: Insufficient organization role, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Conflict: { description: Resource conflict, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    InternalError: { description: Internal server error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    RateLimited: { description: Public Auth rate limit exceeded, headers: { Retry-After: { description: Seconds until another attempt may be made, schema: { type: integer, minimum: 1 } } }, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    ServiceUnavailable: { description: Public Auth protection or a required dependency is unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    PayloadTooLarge: { description: Request body exceeds the endpoint's configured limit (JSON requests are bounded separately from multipart uploads), content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    UnsupportedMediaType: { description: Content-Type must be application/json except for file upload multipart/form-data endpoints, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    ValidationError: { description: Validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
  schemas:
...
    CreateOrganizationRequest: { type: object, additionalProperties: false, required: [name, slug], properties: { name: { type: string, minLength: 2, maxLength: 120 }, slug: { $ref: '#/components/schemas/Slug' } } }
    UpdateOrganizationRequest: { type: object, additionalProperties: false, required: [name, slug], properties: { name: { type: string, minLength: 2, maxLength: 120 }, slug: { $ref: '#/components/schemas/Slug' } } }
...
    CreateProjectUserRequest: { type: object, additionalProperties: false, required: [email, password], properties: { email: { type: string, format: email, maxLength: 320 }, password: { type: string, minLength: 12, maxLength: 256, writeOnly: true }, name: { type: [string, 'null'], minLength: 2, maxLength: 120 } } }
    UpdateProjectUserStatusRequest: { type: object, additionalProperties: false, required: [status], properties: { status: { type: string, enum: [active, blocked] } } }
...
    Slug: { type: string, pattern: '^[a-z0-9][a-z0-9-]{1,62}$' }
...
    Organization: { type: object, required: [id, name, slug, created_at], properties: { id: { type: string, format: uuid }, name: { type: string }, slug: { $ref: '#/components/schemas/Slug' }, created_at: { type: string, format: date-time } } }
    OrganizationPlanLimits: { type: object, additionalProperties: false, required: [projects, members, databases, storage_buckets, functions, sites], properties: { projects: { type: integer, format: int64, minimum: -1 }, members: { type: integer, format: int64, minimum: -1 }, databases: { type: integer, format: int64, minimum: -1 }, storage_buckets: { type: integer, format: int64, minimum: -1 }, functions: { type: integer, format: int64, minimum: -1 }, sites: { type: integer, format: int64, minimum: -1 } }, description: A value of -1 means unlimited. }
    OrganizationPlanUsage: { type: object, additionalProperties: false, required: [projects, members, databases, storage_buckets, functions, sites], properties: { projects: { type: integer, format: int64, minimum: 0 }, members: { type: integer, format: int64, minimum: 0 }, databases: { type: integer, format: int64, minimum: 0 }, storage_buckets: { type: integer, format: int64, minimum: 0 }, functions: { type: integer, format: int64, minimum: 0 }, sites: { type: integer, format: int64, minimum: 0 } } }
    OrganizationPlan: { type: object, additionalProperties: false, required: [organization_id, plan_key, status, current_period_start, current_period_end, limits, usage], properties: { organization_id: { type: string, format: uuid }, plan_key: { type: string, enum: [free, pro, enterprise] }, status: { type: string, enum: [active, past_due, canceled] }, current_period_start: { type: string, format: date }, current_period_end: { type: string, format: date }, limits: { $ref: '#/components/schemas/OrganizationPlanLimits' }, usage: { $ref: '#/components/schemas/OrganizationPlanUsage' } } }
...
    ProjectUser: { type: object, additionalProperties: false, required: [id, project_id, email, name, status, email_verified, created_at, updated_at], properties: { id: { type: string, format: uuid }, project_id: { type: string, format: uuid }, email: { type: string, format: email }, name: { type: [string, 'null'], minLength: 2, maxLength: 120 }, status: { type: string, enum: [active, blocked] }, email_verified: { type: boolean }, created_at: { type: string, format: date-time }, updated_at: { type: string, format: date-time } } }
    Pagination: { type: object, required: [limit, next_cursor], properties: { limit: { type: integer }, next_cursor: { type: [string, 'null'], format: uuid } } }
    ErrorEnvelope: { type: object, additionalProperties: false, required: [error], properties: { error: { type: object, additionalProperties: false, required: [code, message], properties: { code: { type: string }, message: { type: string } } } } }
...
    OrganizationResponse: { type: object, required: [organization], properties: { organization: { $ref: '#/components/schemas/Organization' } } }
    OrganizationPlanResponse: { type: object, additionalProperties: false, required: [plan], properties: { plan: { $ref: '#/components/schemas/OrganizationPlan' } } }
...
    ProjectUserResponse: { type: object, additionalProperties: false, required: [user], properties: { user: { $ref: '#/components/schemas/ProjectUser' } } }
...
    OrganizationsPage: { type: object, required: [organizations, pagination], properties: { organizations: { type: array, items: { $ref: '#/components/schemas/Organization' } }, pagination: { $ref: '#/components/schemas/Pagination' } } }
...
    ProjectUsersPage: { type: object, additionalProperties: false, required: [users, pagination, can_manage], properties: { users: { type: array, items: { $ref: '#/components/schemas/ProjectUser' } }, pagination: { $ref: '#/components/schemas/Pagination' }, can_manage: { type: boolean, description: Whether the current Console membership may create or change project users. } } }
...